Rename "basic" example to "playground", improve playground, and add new basic
