+++ docs/src/examples/components/Tree/playground.svelte
@@ -1,0 +1,160 @@
+<script module lang="ts">
+	import { getFlare } from '$lib/data.remote';
+	let data = await getFlare();
+</script>
+
+<script lang="ts">
+	import type { ComponentProps } from 'svelte';
+	import { cubicOut } from 'svelte/easing';
+	import { hierarchy as d3Hierarchy, type HierarchyNode } from 'd3-hierarchy';
+	import { curveBumpX } from 'd3-shape';
+
+	import { Chart, Group, Link, Layer, Rect, Text, Tree } from 'layerchart';
+	import TransformContextControls from '$lib/components/controls/TransformContextControls.svelte';
+	import TreeControls from '$lib/components/controls/TreeControls.svelte';
+	import { cls } from '@layerstack/tailwind';
+	import type { ConnectorSweep, ConnectorType } from '$lib/utils/connectorUtils.js';
+
+	let config = $state({
+		orientation: 'horizontal' as 'horizontal' | 'vertical' | 'radial',
+		layout: 'chart' as 'chart' | 'node',
+		type: 'd3' as ConnectorType,
+		sweep: 'none' as ConnectorSweep,
+		curve: curveBumpX,
+		radius: 60,
+		siblingGap: 20,
+		parentGap: 100,
+		angularSpacing: 23
+	});
+
+	let expandedNodeNames = $state(['flare']);
+	const hierarchy = $derived(
+		d3Hierarchy(data, (d) => (expandedNodeNames.includes(d.name) ? d.children : null))
+	);
+	// .sum((d) => d.value)
+	// .sort(sortFunc('value', 'desc'));
+	let selected = $state();
+
+	function getNodeKey(node: HierarchyNode<{ name: string }>) {
+		return node.data.name + node.depth;
+	}
+
+	const nodeWidth = 120;
+	const nodeHeight = 20;
+	const nodeSize = $derived(
+		config.orientation === 'radial'
+			? ([(config.angularSpacing * Math.PI) / 180, nodeWidth + config.parentGap] as [
+					number,
+					number
+				])
+			: config.orientation === 'horizontal'
+				? ([nodeHeight + config.siblingGap, nodeWidth + config.parentGap] as [number, number])
+				: ([nodeWidth + config.siblingGap, nodeHeight + config.parentGap] as [number, number])
+	);
+
+	export { data };
+</script>
+
+<TreeControls bind:config />
+
+<Chart
+	padding={config.orientation === 'radial'
+		? 80
+		: { top: 32, left: nodeWidth / 2, right: nodeWidth / 2 }}
+	radial={config.orientation === 'radial'}
+	transform={{
+		mode: 'canvas',
+		motion: { type: 'tween', duration: 800, easing: cubicOut }
+	}}
+	clip
+	height={800}
+>
+	{#snippet children()}
+		<TransformContextControls orientation="horizontal" />
+
+		<Tree
+			{hierarchy}
+			orientation={config.orientation === 'radial' ? 'horizontal' : config.orientation}
+			nodeSize={config.layout === 'node' ? nodeSize : undefined}
+		>
+			{#snippet children({ nodes, links })}
+				<Layer>
+					<Group center={config.orientation === 'radial'}>
+						<Group opacity={0.2}>
+							{#each links as link (getNodeKey(link.source) + '_' + getNodeKey(link.target))}
+								<Link
+									data={link}
+									orientation={config.orientation === 'radial' ? undefined : config.orientation}
+									curve={config.curve}
+									type={config.type}
+									sweep={config.sweep}
+									radius={config.radius}
+									motion="tween"
+									class="stroke-surface-content"
+								/>
+							{/each}
+						</Group>
+
+						{#each nodes as node (getNodeKey(node))}
+							{@const nodeX =
+								config.orientation === 'radial'
+									? node.y * Math.sin(node.x)
+									: config.orientation === 'horizontal'
+										? node.y
+										: node.x}
+							{@const nodeY =
+								config.orientation === 'radial'
+									? -node.y * Math.cos(node.x)
+									: config.orientation === 'horizontal'
+										? node.x
+										: node.y}
+							<Group
+								x={nodeX - nodeWidth / 2}
+								y={nodeY - nodeHeight / 2}
+								motion="tween"
+								onclick={() => {
+									if (expandedNodeNames.includes(node.data.name)) {
+										expandedNodeNames = expandedNodeNames.filter((name) => name !== node.data.name);
+									} else {
+										expandedNodeNames = [...expandedNodeNames, node.data.name];
+									}
+									selected = node;
+
+									// transform.zoomTo({
+									//   x: orientation === 'horizontal' ? selected.y : selected.x,
+									//   y: orientation === 'horizontal' ? selected.x : selected.y,
+									// });
+								}}
+								class={cls(node.data.children && 'cursor-pointer')}
+							>
+								<Rect
+									width={nodeWidth}
+									height={nodeHeight}
+									class={cls(
+										'fill-surface-100',
+										node.data.children
+											? 'stroke-primary hover:stroke-2'
+											: 'stroke-secondary [stroke-dasharray:1]'
+									)}
+									rx={10}
+								/>
+								<Text
+									value={node.data.name}
+									x={nodeWidth / 2}
+									y={nodeHeight / 2}
+									dy={-2}
+									textAnchor="middle"
+									verticalAnchor="middle"
+									class={cls(
+										'text-xs pointer-events-none',
+										node.data.children ? 'fill-primary' : 'fill-secondary'
+									)}
+								/>
+							</Group>
+						{/each}
+					</Group>
+				</Layer>
+			{/snippet}
+		</Tree>
+	{/snippet}
+</Chart>
